swift_ai_pipeline_fastapi/requirements.txt
fastapi
uvicorn
openai
pdfplumber
pandas
python-multipart
pytest
python-dotenv
pytest-asyncio
pymupdf

